Altered domain for road standard vehicles
The different vehicle types are now condensed into one StandardVehicles object. Each vehicle type can have it's own vehicles-per-time-unit and stagnation factor, but they are all assumed to have the same time unit and maximum speed.
This approach works better for the UI, but it now is inconsistent with GML, so converting to/from those objects has been adjusted. It also means that the base object no longer has a shared method to obtain number of vehicles, as this depends on the specific type.
Not a huge fan of this approach, but if it works...
Same as changes in IMAER part in commit 82d1779a5d47c99379e98cfeb7a950037d43922c on main AERIUS repo