AER3-1097 Added generic emissions per subsource
These emissions are calculated by the different emission calculators (and set in the supplied objects by those emission calculators as a side-effect of determining the total emission).
It does muddle the domain objects a bit in the sense that it might not be directly obvious that the emissions for the subsource are calculated based on the properties, or is input by the user (in the case of farmland).