Added CalculationOption to java GML generation
Using an enum to go from CalculationSetOptions to a map with key-value pairs. This method can be used both in GML (where default values are not exported) and in other parts (where defaults might be added anyway).
Removed monitorSrm2Year from being readable: it was the only calculation metadata field that was readable that way, and we weren't using this functionality anyway except for testing.
Also exporting monitorSrm2Year with the new functionality. It was a pretty odd one anyway, only required within the NSL context. In the XSD it's still available as a specific property, but with this change we won't ever export it that way (making it deprecated).