AER3-1040 generic calculation options (#10)
* Changed IMAER to contain (optional) generic calculation options
That is, these options can be added to the calculation metadata to show which options were used for calculation results.

* Added CalculationOption to java GML generation
Using an enum to go from CalculationSetOptions to a map with key-value pairs. This method can be used both in GML (where default values are not exported) and in other parts (where defaults might be added anyway).
Removed monitorSrm2Year from being readable: it was the only calculation metadata field that was readable that way, and we weren't using this functionality anyway except for testing.
Also exporting monitorSrm2Year with the new functionality. It was a pretty odd one anyway, only required within the NSL context. In the XSD it's still available as a specific property, but with this change we won't ever export it that way (making it deprecated).

* Changed XSD version to 4.0.2
Since 4.0.1 has been released, seems prudent to change the version to 4.0.2. The changes so far shouldn't break validity of a 4.0.1 exported file with this 4.0.2 version (only added optional fields).