Changed the structure that Meteo uses (#11)
That is, it's no longer abstract, but incorporates both start and end year. This way it replaces SingleYearMeteo and MultiYearMeteo. If they are the same, it can be considered a single year meteo, but that shouldn't matter too much for most applications. Removed the opsFile to be part of the object as well: this is too model specific, and determining the correct files is something that can be done in the OPS part of AERIUS based on the start and end year (there's a certain structure to those file names).
This is in preparation of using meteo for ADMS and making it selectable in the UI.
Using @deprecated to avoid causing compilation errors in the Calculator repository.