Introduced Theme to Scenario (#14)
Because we get more themes that have similar calculation types it makes sense to remove the theme like types from CalculationType and add the Theme as a separate variable to Scenario. This will make it easier to use the same calculation type among different themes.
This change is backward compatible, and will be need an additional change once upstream has been updated to work with this change.