AER3-1224 Shipping mooring backwards comp. route names (#22)
* AER3-1224 Constructing labels for mooring routes for inland shipping
There are no real names for routes in the old GML version, so we had to come up with something based on an index.
Took the liberty of letting GML2InlandMooringForceWaterway extend the 'normal' 3.1 version to avoid some duplicate code. Did need some protected methods for that, but since it's backwards compatibility anyway I don't think that matters too much.

* AER3-1224 Constructing labels for mooring routes for maritime shipping
Similar to inland version, but the maritime routes ('buitengaats') were more or less global in old Calculator, so did not include the mooring source name in those labels and using a 'global' index.