Refactored GML2SourceCharacteristics to be more generic
Created an interface and OPS specific implementation. This will make it possible to add other type of characteristics, like ADMS, to the GML.
For older GML versions (everything < 4.0) only OPS is supported therefor it uses always OPS characteristics.
For 4.0 it now only supports OPS, but contains code a basis to add more (it's left open if this will be added to version 4.0 or a new version is needed. For this implementation that is not relevant).