Switched to newer JTS library (1.18.2)
In 2016/2017 JTS moved to locationtech. With version 1.15 (nov 2017) the package names and maven groupIds were changed as well. Think now is a good time to switch along with it (gives some more options as well, which might help with the overlapping hexagon detection for instance). Should contain some performance improvements as well (though not too sure how that'll affect us).
This will probably break applications when using classes like GeometryUtil.