Use an ImaerImporter that doesn't use the static GMLReaderFactory
The static GMLReaderFactory gets a mocked GMLHelper.
On the GitHub server the tests seem to run in 1 jvm run and therefor the
same GMLReaderFactory is returned. However, because the GMLHelper is
mocked it's a different one in different tests.
But the initial mocked GMLHelper doesn't work well in the other tests
causing the mocking to return empty in other tests.