AER-34 road changes for IMAER 5.0 (#53)
* Adjusted imaer for new road structure
No longer have speed profiles, instead use road type (on the base class so it affects both SRM1 and SRM2). VehicleType no longer an enum, now a CharacterString (should be a code).
* Checking on exceptions before doing anything else
This makes some more sense, as now any XSD error will be shown instead of an error caused due to not reading the file (like not finding any sources or the correct version).
* Adjusted domain object for new road structure
* Adjusted imaer-util for new domain structure
* Moved some GML2 classes to subpackages
* Properly writing IMAER 5.0
* Moved road related classes to v40 package
Also made the necessary changes for the new domain structure, validator method and emission factor supplier methods.
* Schematron test files to 5.0
Annoying that this is needed. Still want to get rid of schematron, it doesn't add anything.
* Changes on v5_0 test files to follow new XSD
Changed all sectors to 3100 for road, as the current idea is to get rid of the sectors-as-road-type difference and just have 1 sector for roads.
* Implemented road changes for imaer 5.0
* Changed IMAER-java version to 5.0-SNAPSHOT
With these changes we're actually exporting 5.0, so makes sense to change the version of the jars. Also helps with compatibility in other projects...