AER-324 interpolation for ADMS road emission factors (#65)
* Split SRM Road and ADMS Road emission calculations
ADMS won't use stuff like stagnation factor, use a different target time unit and will use interpolation to determine emission factors. Easier to do that in a separate class...
* Introduced a key class to reduce the number of arguments
Should also reduce chance of switching arguments around (though there's still a chance when making the actual key). As it was, the validation method used a different order of the first 3 arguments compared to the methods obtaining emission factors.
* Interpolation for road for speed and gradient
Tried to keep the interpolation technique in IMAER-java, though obtaining the actual values between which to interpolate is something to be done in the code using these calculators. Another option would be to ask for lists of possible values for speed and gradient, given the roadtype/vehicletype/etc.
* Adjusted test class for new method
Might adjust implementation later on, but this should work for now.
* Added messages to asserts in unittests