Move theme specific calculation options to it's own class
In CalculationSetOptions different theme options are present.
This change adds new classes with the specific options.
For now all available options are present in the CalculationSetOptions to make handling the code not overly complex in the existing code base.

This change also adds new ADMS calculation options.
To be backward compatible old methods are deprecated.