Move theme specific calculation options to it's own class (#74)
In CalculationSetOptions different theme options are present. This change adds new classes with the specific options. Added theme to MetaDataInput to distinguish between different calculation options. If Theme not set it defaults to WNB to be backward compatible.

All available options are present in the CalculationSetOptions to make handling the code not overly complex in the existing code base. Making them on available depending on context requires either lots of generics or lots of checks (to see if options are present).

This change also adds new ADMS calculation options.
To be backward compatible old methods in CalculationSetOptions are deprecated.