AER-353 Overlapping property in IMAER 4.0
Includes changing version from 4.0.2 to 4.0.3.
Includes a unittest that shows how it works (though for the test to trigger you have to manually switch to 4.0 in InternalGmlWriter...)
Did add a test case for 5.0, but didn't add the property yet as I'd rather have feedback on this before adding it to 5.0.