Diurnal variation for ADMSRoadEmissionSource
Added diurnal variation in IMAER and IMAER-objects earlier, but the actual domain part was missing.
Similar to the approach for source characteristics at the moment, there either is no definition (in which case it should use the default), there is a standard DV code, or there is a custom definition.