AER-361 Deprecated massflux property in IMAER-java (#119)
Never added this property to actual IMAER, so any GML that would be exported with the massFlux property would be invalid w.r.t. the XSD. As we're not supporting the mass efflux type any more, we might as well get rid of the property while we're still in 5.0 territory.