Generic validation for FarmEmissionFactorType
The 'expectsNumberOfAnimals' and 'expectsNumberOfDays' methods in FarmlandValidationHelper didn't make too much sense in hindsight: the emission factor type is already something known in this domain model, so might as well ask the helper for the specific type for a category code.
The new interface is an attempt to reduce code duplication and to be ready for the new emission factor types to be actually used.