Introduced a CalculationPointDuplicator
Used the CalculationPointVisitor, that has been updated with the recently added types, so using `calculationPointFeature.visit(duplicator)` should be enough to duplicate a point (without it's results).
Methods are protected with the idea that this should make it easier to override any of the types, in which case this class should be extended for the appropriate type (no longer visitor pattern in that case, but wouldn't know how else to do that with types not available in this library).