AER-1065 duplicate coordinate check (#158)
* Test case for duplicate consecutive points in polygon WKT definition
We want these to be valid for intersections.
* Using deprecated IsSimpleOp
This was changed with changing JTS version: we moved from the deprecated version to the non-deprecated version. However, implementation is different:
'Polygonal geometries are simple by definition, so isSimple trivially returns true' in the old version, whereas the new version checks for 'simple polygon rings', which in turn checks for not self-intersection at interior points which includes duplicate coordinates (apparantly).
See the following links:
https://locationtech.github.io/jts/javadoc/org/locationtech/jts/operation/IsSimpleOp.html
https://locationtech.github.io/jts/javadoc/org/locationtech/jts/operation/valid/IsSimpleOp.html
Changing back to the deprecated version means a succesful test case at least.
* Skip the simple check altogether
The IsSimple approach was used to check if linestrings were self-intersecting. As we're specifically testing polygons only, that doesn't need to be done (it doesn't help with checking for intersections within polygons).