AER-941 GML coherence
Or rather, including information about other situations that were included in a calculation when including results.
In AERIUS NL, the points to calculate are actually influenced by other situations, so it helps if it is discernable from a GML that other situations were around when calculating.
While the IMAER definition includes a reference field for 'OtherSituationMetadata', this won't be exported (yet): currently the reference is always generated during GML generation, and the reference is supplied through the ScenarioMetaData, which would be the same for all situations... So if that's a critical requirement, that requires some more work.