AER-1377 Introduced CalculationMethod as replacement for CalculationType
CalculationMethod will be the new name for CalculationType.
Older GML files will still have the calculation type, but when read are migrated to CalculationMethod.
For version IMAER 5.1 calculation method will be written in the GML.
For version IMAER 5.0 will be written in the GML.
In Calculation Method the method RADIUS has been removed as it's not used anymore.

TODO: Update the xsd to support calculation method. It will be an addition.
(Although 'type' changes that it isn't required anymore)
In this change the test gml files are already adapted to use 'method' element.

Possible TODO: rename PERMIT to FORMAL to align it better with the general usage and UK.
(The code already contains some specific conversion in case FORMAL will be introduced).

This is a breaking change in the Java library.