Reading/writing archive metadata in java
Including domain objects to be able to do a roundtrip test (need those at some point anyway).
We could introduce a fully new object instead of abusing MetaDataInput I guess, to be more strict on what we expect in an Archive GML. But for now this seems OK, it should at least be enough to get the export going.