Using empty GWT wrapper class for java.time.ZonedDateTime (#226)
ArchiveMetaData contains ZonedDateTime, and that is not emulated by GWT apparantly. To avoid having to resort to some other data type(date or string perhaps?), we might as well use an empty wrapper similar to what we've been using for LocalDate, as these shouldn't be serialized to GWT code anyway.
Also added some Serializable interface usage.