AER-2477 Validate by schema when writing GML (#240)
* Validate by schema when writing GML
By setting the schema on the marshaller, any XSD violations should be caught before invalid GMLs end up at the user.
* Review suggestion: specifying schema locations in version enum
* Removed (internal) schema location from CalculatorSchema
As this location is only necessary for reading the schema in code, might as well do it like this. Other CalculatorSchema stuff is referenced in annotations, so probably better to keep that like it is.