Changes due to using Shapechange 3.0.0
Moved from 2.0.1 (which we have been using from the start) to 3.0.0 (which was released last year).
Did mean all the documenation in the EAP file had to be updated to remove `-- Definition --` and `-- Description --` as those are no longer treated as a special case, but that was not really adding anything anyway.
Also got rid of HTML that was added to version control, but which is actually generated during shapechange run.
Also updated instructions and changed that to markdown.