AER-2750 Shapechange 3.0.0 and cleanup (#252)
* Changes due to using Shapechange 3.0.0
Moved from 2.0.1 (which we have been using from the start) to 3.0.0 (which was released last year).
Did mean all the documenation in the EAP file had to be updated to remove `-- Definition --` and `-- Description --` as those are no longer treated as a special case, but that was not really adding anything anyway.
Also got rid of HTML that was added to version control, but which is actually generated during shapechange run.
Also updated instructions and changed that to markdown.
* Renaming time varying profiles
Also including monthly profiles for ADMS road in addition of hourly profiles (overlooked when adding it in ADMS source characteristics that this was also a property on ADMSRoad).
Also got rid of some deprecated proprerties.
* Removing Plan emission sources
Hasn't been used for a while, so time to finally remove these.
* Changed last occurence of diurnal variation
At this point we might as well, as in IMAER OPS characteristics also use the time varying profile approach (though only the standard one, not custom).
* Got rid of some metadata fields that were no longer used
Also removed CalculationType enum from model, as it is no longer used.
* XMI version of EAP file
Not sure yet if this will be beneficial. I guess only if there is some discipline to updating this file when changing something in the EAP file.