Adjusted setting description for custom emission
This is only used when conversion was not present, which shouldn't be all that often. But if it does, then the empty string is a better indicator, as that's not something that you can easily do through the UI. And do need something, otherwise resulting GML would be invalid (it's not an optional field).