AER-2858 farm lodging backwards compatibility (#267)
* Farm lodging conversion to Farm animal housing
If no conversion data around at all, the old farm lodging conversion happens, which should work as a sort of feature flag for now.
* Testing changes
* Proper reasons for warning when converting lodgings
Also using empty string as unknown animal type code, which can be used for some custom source conversions.
* Removed old conversion files
Not entirely sure why these were still around, but tests succeed without these so don't really see a point in keeping them. Also, 0.5, how long has that been deprecated by now?
Removing the road ones does have an effect, so letting those be for now.
* Adjusted setting description for custom emission
This is only used when conversion was not present, which shouldn't be all that often. But if it does, then the empty string is a better indicator, as that's not something that you can easily do through the UI. And do need something, otherwise resulting GML would be invalid (it's not an optional field).
* Avoid warning when code did not change
This is mostly for UK, where we don't want warnings because nothing really changed.