Adjusted conversion from/to gml objects to include new fields
Had to add some util classes and thrown exceptions for the centroid conversion.
Also had to update the test case, as I went with a slightly different prefix for the ID (to avoid any conflict with existing 2-letter prefixes) and the XML formatted generation still has this issue when indentation becomes too much to reset that indentation.