AER-3099 Added support for having multi adms met site properties for multiple years. (#303)
This code introduces a new MetSurfaceCharacteristics object that replaces the single values in ADMSOptions.
For backward compatibility the individual methods are kept in ADMSOptions, and will be removed when not used any more by upstream project. Also added MetSiteLatitude and WindInSectors as these are also relevant ADMS options.
Made the 'old' MsMetxx' methods deprecated and JsonIgnore to avoid serialisation problems.