Renamed some non-adms metadata fields
These fields are not used in the ADMS model, and do not have any effect on input there, so not much point in having it prefixed with adms_. Could use nca_, but as they're generic codes anyway, we might as well use generic names in the metadata.
Yes, this does 'break' old imports, but now seems a good moment to do this as we're moving to a new beta release.