AER-3433 keep track of calculated snapshot values
Reasoning for putting it in CalculationSetOptions and the generic metadata section in IMAER is that its least changes. Conceptually it makes less sense as it's not a value that a user should generally alter/supply. But that's similar to OPS version that's also present in options/metadata I guess?
Not too sure about both the reasoning and the term CalculatedSnapshotValues, suggestions are welcome.