Switched to ogc-schemas jar
The old jar that was used, was also used for jaxb classes. However, that project has underwent significant changes (being forked to another repo for one), and schemas are now available in a separate jar. That one contains a lot of OGC xsd stuff, so using some more specific filtering. Also the xsds no longer have relevant locations, so had to put some more stuff in gml-catalog, and including some w3.org xsds in our repository. Was able to get rid of the gmlsProfile directory though, which is a plus (and this library is not 12 years old anymore).