Update dependencies (#314)
* Updating dependencies in pom.xmls
* Java changes for new dependencies
Mostly javax -> jakarta, but also net.opengis.gml.v_3_2_1 -> net.opengis.gml_v_3_2 (which is newer, don't ask)
* Switched to ogc-schemas jar
The old jar that was used, was also used for jaxb classes. However, that project has underwent significant changes (being forked to another repo for one), and schemas are now available in a separate jar. That one contains a lot of OGC xsd stuff, so using some more specific filtering. Also the xsds no longer have relevant locations, so had to put some more stuff in gml-catalog, and including some w3.org xsds in our repository. Was able to get rid of the gmlsProfile directory though, which is a plus (and this library is not 12 years old anymore).
* Moved some version properties to the parent pom
A bit easier to see at a glance what versions are used if they're centralized.
Did skip a few in imaer-server-xsd-standalone, as those shouldn't be used outside of it anyway.
* Not updating JTS due to change in results
Apparantly with the new version something changed so length of geometries that is determined in JTS changes slightly. At a glance this is rounding, but better to make a decision on this if it's OK or not.
* Specify java version for shared modules
With GWT 2.12 it should be OK to use java 17 stuff. If we have to go back to 2.10 for downstream reasons, we can put it back to 11. Main reason is to avoid using constructs that are not available in GWT.