AER-3516 Implemented reading spread with legacy support
In new OPS version spread for points should be passed by the user.
This means if no spread was given it should be getting half the emission height.
In GML the spread can be present, but for point and lines it was not specified what the value meant as it was not used.
Therefore we should treat the spread value in older GML's as a random value, and always derive the spread from the emission height.
This change adds support for either reading the spread or deriving it from emission height (legacy).
For now all versions of GML will use the legacy variant, as all versions are also available in production.

As of this moment there is not support for writing the spread to the GML for point/line sources,
as it will not be possible to determine if it was legacy or set correctly since the GML version is not changed,
and there are already GMLs generated in production with the legacy situation.
To be able to correctly set the spread some new version of IMAER must be made.