AER-3735 default characteristics also determined by geometry type (#335)
For ADMS it matters what the geometry type is. Or rather, it matters what the ADMS source type is, but the default for that can be based on geometry type. And since geometry type is fairly generic...
Default methods in GMLCharacteristicsSupplier for easier conversion (that's the idea behind it at least).