AER-3576 Added reading/writing spread for point/line sources (#340)
From IMAER v6.0 created with AERIUS version 2025 or newer the source spread will be read from the IMAER file. For all previous IMAER version and V6.0 created with AERIUS version 2024 the spread for point and line sources will be calculated from the emission height and any spread in the IMAER file will be ignored.