AER-3576 Reimplemented handling spread for point/line sources (#342)
For point/line sources OPS did not use the spread. i.e. it was interpreted as being 0. Therefore for backward compatibility the spread for point/line sources for older versions of IMAER, and latest (6.0) created with AERIUS 2024 the spread will be set to 0 for point/line sources. If the spread is missing from a newer GML the spread for point/line sources will be set to 0. This only will be a case if the user manually modified the GML, because the spread will always be present in a new GML.