AER-3888 Force gml id when reading GML.
This is to set gml on older maritime routes migrated to newer version in which they are separate emission sources.
Before they didn't get a gmlId. Theoretically this is not a problem because when writing a GML they get an gml id anyway.
But to make code more consistent they also get an gml id when reading.
This will also solve issue where old maritime mooring with route could be imported, but not being imported because the sources with null gml id's are registered as being duplicated.
The code was also refactored to make setting gml ids more generic and moved out version specific writers.