AER-3988 Improve flow on handling import year (#350)
When importing the sources are checked for having emissions. However when both the specific import year and year in GML are for years for which no emission factors are available all sources that calculate emissions will be reported as having no emissions. This is not desired because upstream the year will be set to an valid year and emissions will be recalculated again making these earlier reported missing emissions invalid. Therefore it is better to always make sure the year used is a valid year. Because it's not known what a valid year is this is delegated by adding a method to GMLHelper to get the year. The implementation should always return a valid year, and also the implementation has the option to report a warning about the year in case it was adjusted.

This implementation uses a default implementation in the interface, which makes it backwards compatible. With this new implementation we always assume the year is adjusted to a valid year it doesn't add much to also validate the year to be out of certain ranges. Therefore these GMLValidator#validateYear method has been removed.